chore(plugin): remove RelationColumnSet feature (unused, superseded by exo__Layout)
User directive 2026-06-15: "Удали relation-column-set-smoke тест И фичу,
которую он тестирует — я ей не пользуюсь." RelationColumnSet (ui__RelationColumnSet,
RFC be70f741) is a legacy layout feature superseded by exo__Layout.

Clean removal — RelationColumnSet was an optional, feature-flagged parallel
layer with an explicit fallback in RelationsRenderer; exo__Layout rendering is
unaffected. No CLI migrator existed (already removed in RFC 78c2b7d0).

Removed:
- exocortex: domain/layout/RelationColumnSet + application/services/RelationColumnSetResolver.
  Relocated the shared `normalizeRef` helper to domain/layout/normalizeRef.ts — still
  used by Layout/LayoutBlock frontmatter parsers. Barrel exports dropped.
- obsidian-plugin: RelationColumnSetRepository + ObsidianRelationColumnSetAdapter;
  RelationsRenderer feature branch (now always returns the legacy hardcoded group map);
  UniversalLayoutRenderer wiring; ExocortexPlugin onload/dispose wiring; the
  enableRelationColumnSetResolver setting + VaultSettingsRegistry entry.
- exoas-exo submodule: exo__SettingKeyEnableRelationColumnSetResolver SettingKey
  individual (keeps VaultSettingsRegistry<->graph parity 1:1); submodule pointer bumped.
- All RelationColumnSet tests + fixtures + e2e test-vault + the relation-column-set-smoke
  spec (also removed from playwright-shard-assignments.json).
- Docs: RELATION_COLUMN_SET.md, EXO_LAYOUT.md coexistence/migration sections, README row.

Verified locally: build green (tsc --noEmit + esbuild + mobile-loadable + console-channel),
lint clean, exocortex 8037 tests pass, plugin settings count+parity 24<->24, shard
validator OK.

Vault ui__RelationColumnSet TBox class + properties + instances cleanup is deferred to a
separate vault-cleanup task (does not gate this code PR).
