+++ packages/exocortex/README.md
@@ -16,18 +16,18 @@ npm run build -w exocortex
 
 ## Subsystems
 
-| Subsystem | Where | What it does |
-| --- | --- | --- |
-| RDF core | `src/domain/models/rdf/`, `src/infrastructure/rdf/` | `IRI` / `Literal` / `BlankNode` / `Triple` (incl. RDF-Star `QuotedTriple`), `InMemoryTripleStore` (6-way indexed, named graphs, UUID index), `RDFSerializer`, `RDFSInferenceEngine` |
-| SPARQL / ExoQL engine | `src/infrastructure/sparql/`, `src/exoql/` | Parser → algebra translator → executors pipeline; `ExoQL` facade (`query` / `ask` / `construct`); aggregates, query-plan & result caches, optimizers |
-| Vault → RDF conversion | `src/services/NoteToRDFConverter.ts` | Converts markdown frontmatter into triples; folder exclusions, invariant-violation reporting |
-| SHACL-lite validation | `src/services/ShaclLiteValidator.ts`, `ShapeLoader`, `ShapeRegistry`, `ValidatorDaemon` | Validates triples against shapes (required properties, cardinality, class membership with subclass closure) |
-| Homoiconic commands | `src/services/CommandResolver.ts`, `PreconditionEvaluator`, `GroundingExecutor`, `CommandExecutionFlow` | Resolves `exocmd__Command` / `Precondition` / `Grounding` assets declared in the vault into executable UI commands |
-| Sync (ExoSync) | `src/services/sync/`, `src/infrastructure/github/` | `detectChanges` + `SyncEngine` (pull → merge → push over the `restCreateCommit` Git Data API core), `StructuredMerger` (3-way frontmatter/body merge), `MergeShaclGate`, quarantine & watermark stores |
-| AssetSpace & profile | `src/services/assetspace/`, `src/domain/profile/`, `src/services/profile/` | `mountAssetSpaceFiles` mount pipeline + `.gitmodules` transforms, tarball parser, `assertTsFloor` floor guard, `IConfirmGate` apply-plan contract |
-| Asset & status services | `src/services/` | `TaskStatusService`, `PlanningService`, `EffortStatusWorkflow`, `GenericAssetCreationService`, `ArchiveAssetService`, `FolderRepairService`, `RenameToUidService`, and more |
-| Layout | `src/domain/layout/`, `src/application/services/` | `exo__Layout` / `LayoutBlock` / `RelationColumnSet` models, `LayoutSelector`, `RelationColumnSetResolver` |
-| Utilities & DI | `src/utilities/`, `src/interfaces/`, `src/infrastructure/container.ts` | `FrontmatterService`, `DateFormatter`, `WikiLinkHelpers`, `MetadataHelpers`, `FilenameValidator`; adapter interfaces and the DI container |
+| Subsystem               | Where                                                                                                   | What it does                                                                                                                                                                                           |
+| ----------------------- | ------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| RDF core                | `src/domain/models/rdf/`, `src/infrastructure/rdf/`                                                     | `IRI` / `Literal` / `BlankNode` / `Triple` (incl. RDF-Star `QuotedTriple`), `InMemoryTripleStore` (6-way indexed, named graphs, UUID index), `RDFSerializer`, `RDFSInferenceEngine`                    |
+| SPARQL / ExoQL engine   | `src/infrastructure/sparql/`, `src/exoql/`                                                              | Parser → algebra translator → executors pipeline; `ExoQL` facade (`query` / `ask` / `construct`); aggregates, query-plan & result caches, optimizers                                                   |
+| Vault → RDF conversion  | `src/services/NoteToRDFConverter.ts`                                                                    | Converts markdown frontmatter into triples; folder exclusions, invariant-violation reporting                                                                                                           |
+| SHACL-lite validation   | `src/services/ShaclLiteValidator.ts`, `ShapeLoader`, `ShapeRegistry`, `ValidatorDaemon`                 | Validates triples against shapes (required properties, cardinality, class membership with subclass closure)                                                                                            |
+| Homoiconic commands     | `src/services/CommandResolver.ts`, `PreconditionEvaluator`, `GroundingExecutor`, `CommandExecutionFlow` | Resolves `exocmd__Command` / `Precondition` / `Grounding` assets declared in the vault into executable UI commands                                                                                     |
+| Sync (ExoSync)          | `src/services/sync/`, `src/infrastructure/github/`                                                      | `detectChanges` + `SyncEngine` (pull → merge → push over the `restCreateCommit` Git Data API core), `StructuredMerger` (3-way frontmatter/body merge), `MergeShaclGate`, quarantine & watermark stores |
+| AssetSpace & profile    | `src/services/assetspace/`, `src/domain/profile/`, `src/services/profile/`                              | `mountAssetSpaceFiles` mount pipeline + `.gitmodules` transforms, tarball parser, `assertTsFloor` floor guard, `IConfirmGate` apply-plan contract                                                      |
+| Asset & status services | `src/services/`                                                                                         | `TaskStatusService`, `PlanningService`, `EffortStatusWorkflow`, `GenericAssetCreationService`, `ArchiveAssetService`, `FolderRepairService`, `RenameToUidService`, and more                            |
+| Layout                  | `src/domain/layout/`, `src/application/services/`                                                       | `exo__Layout` / `LayoutBlock` models, `LayoutSelector`                                                                                                                                                 |
+| Utilities & DI          | `src/utilities/`, `src/interfaces/`, `src/infrastructure/container.ts`                                  | `FrontmatterService`, `DateFormatter`, `WikiLinkHelpers`, `MetadataHelpers`, `FilenameValidator`; adapter interfaces and the DI container                                                              |
 
 Detailed API documentation lives in [`docs/api/Core-API.md`](../../docs/api/Core-API.md).
 
@@ -83,10 +83,10 @@ const removed = fm.removeProperty(updated, "ems__Effort_status");
 
 ## Consumers
 
-| Package | How it uses this package |
-| --- | --- |
+| Package                                             | How it uses this package                                                                                                    |
+| --------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------- |
 | [`@exocortex/obsidian-plugin`](../obsidian-plugin/) | Implements `IVaultAdapter` over the Obsidian API; renders layouts, command buttons, and SPARQL blocks driven by this engine |
-| [`@kitelev/exocortex-cli`](../cli/) | Implements the adapters over Node `fs`; exposes `query`, `validate schema`, `apply`, `audit`, and other commands |
+| [`@kitelev/exocortex-cli`](../cli/)                 | Implements the adapters over Node `fs`; exposes `query`, `validate schema`, `apply`, `audit`, and other commands            |
 
 ## Development
