--- packages/exocortex/src/application/services/index.ts
@@ -1,6 +1,0 @@
-export {
-  RelationColumnSetResolver,
-  type RelationColumnSetProvider,
-  type RelationColumnSetResolverLogger,
-  type RelationColumnSetResolverOptions,
-} from "./RelationColumnSetResolver";
