+++ packages/exocortex/src/domain/layout/index.ts
@@ -1,13 +1,4 @@
-export {
-  RELATION_COLUMN_SET_CLASS_IRI,
-  RELATION_COLUMN_SET_CLASS_UID,
-  createRelationColumnSetFromFrontmatter,
-  isRelationColumnSet,
-  isRelationColumnSetFrontmatter,
-  normalizeRef,
-  type CreateRelationColumnSetOptions,
-  type RelationColumnSet,
-} from "./RelationColumnSet";
+export { normalizeRef } from "./normalizeRef";
 
 export {
   LAYOUT_CLASS_IRI,
