+++ packages/obsidian-plugin/playwright-shard-assignments.json
@@ -9,8 +9,7 @@
     ["effort-timestamps-auto-sync.spec.ts", "daily-archive-filter.spec.ts"],
     [
       "dynamic-command-buttons-render.spec.ts",
-      "alias-sync-on-label-change.spec.ts",
-      "relation-column-set-smoke.spec.ts"
+      "alias-sync-on-label-change.spec.ts"
     ],
     [
       "layout-overflow-styling.spec.ts",
