+++ packages/obsidian-plugin/src/infrastructure/repositories/index.ts
@@ -1,13 +1,4 @@
 export {
-  RelationColumnSetRepository,
-  type RelationColumnSetEventHandler,
-  type RelationColumnSetLogger,
-  type RelationColumnSetRepositoryOptions,
-  type RelationColumnSetSnapshot,
-  type RelationColumnSetVaultAdapter,
-} from "./RelationColumnSetRepository";
-export { ObsidianRelationColumnSetAdapter } from "./ObsidianRelationColumnSetAdapter";
-export {
   ExoLayoutRepository,
   type ExoLayoutEventHandler,
   type ExoLayoutLogger,
