--- packages/obsidian-plugin/tests/e2e/test-vault/relcolset/week-2026-w17.md
@@ -1,12 +1,0 @@
----
-exo__Instance_class: "[[period__Week]]"
-exo__Asset_uid: fixture-week-2026-w17
-exo__Asset_label: "Week 2026-W17 (RCS fixture)"
-exo__Asset_createdAt: "2026-04-20T00:00:00+0500"
----
-
-# Week 2026-W17
-
-This asset is the row-parent for the RelationColumnSet Phase 3 E2E smoke spec.
-Three `ems__WeeklyObjective` notes reference this file through
-`ems__WeeklyObjective__week`.
