--- packages/obsidian-plugin/tests/e2e/test-vault/relcolset/wo-1.md
@@ -1,9 +1,0 @@
----
-exo__Instance_class: "[[ems__WeeklyObjective]]"
-exo__Asset_uid: fixture-wo-1
-exo__Asset_label: "Ship Phase 3 integration"
-exo__Asset_createdAt: "2026-04-20T09:00:00+0500"
-ems__WeeklyObjective__week: "[[week-2026-w17]]"
----
-
-# Ship Phase 3 integration
