--- packages/obsidian-plugin/tests/e2e/test-vault/relcolset/wo-2.md
@@ -1,9 +1,0 @@
----
-exo__Instance_class: "[[ems__WeeklyObjective]]"
-exo__Asset_uid: fixture-wo-2
-exo__Asset_label: "Close CI Speedup Phase 4"
-exo__Asset_createdAt: "2026-04-20T10:00:00+0500"
-ems__WeeklyObjective__week: "[[week-2026-w17]]"
----
-
-# Close CI Speedup Phase 4
