--- packages/obsidian-plugin/tests/e2e/test-vault/relcolset/wo-3.md
@@ -1,9 +1,0 @@
----
-exo__Instance_class: "[[ems__WeeklyObjective]]"
-exo__Asset_uid: fixture-wo-3
-exo__Asset_label: "Vault audit backlog"
-exo__Asset_createdAt: "2026-04-20T11:00:00+0500"
-ems__WeeklyObjective__week: "[[week-2026-w17]]"
----
-
-# Vault audit backlog
